div 1000000

+++ eve_parser/scripts/logistics_planning.py
@@ -39,8 +39,8 @@ def calculate_logistics(args):
                     price_from=liquidity_from.price, price_to=liquidity_to.price, price_diff=liquidity_from.price - liquidity_to.price,
                     liquidity_from=liquidity_from.day_turnover, liquidity_to=liquidity_to.day_turnover,
                     day_volume_from=liquidity_from.day_volume, day_volume_to=liquidity_to.day_volume,
-                    profit_from=(liquidity_from.price - liquidity_to.price) * (-1) * liquidity_to.day_volume / 1000000,
-                    profit_to=(liquidity_from.price - liquidity_to.price) * liquidity_from.day_volume) / 1000000
+                    profit_from=float((liquidity_from.price - liquidity_to.price) * (-1) * liquidity_to.day_volume) / 1000000,
+                    profit_to=float((liquidity_from.price - liquidity_to.price) * liquidity_from.day_volume)) / 1000000
                 logistics_planning.save()
             else:
                 logistics_planning = LogisticsPlanning.objects.filter(type_id=item_type[0], region_id_from=region_from,
@@ -49,8 +49,8 @@ def calculate_logistics(args):
                     price_diff=liquidity_from.price - liquidity_to.price,
                     liquidity_from=liquidity_from.day_turnover, liquidity_to=liquidity_to.day_turnover,
                     day_volume_from=liquidity_from.day_volume, day_volume_to=liquidity_to.day_volume,
-                    profit_from=(liquidity_from.price - liquidity_to.price) * (-1) * liquidity_to.day_volume / 1000000,
-                    profit_to=(liquidity_from.price - liquidity_to.price) * liquidity_from.day_volume / 1000000)
+                    profit_from=float((liquidity_from.price - liquidity_to.price) * (-1) * liquidity_to.day_volume) / 1000000,
+                    profit_to=float((liquidity_from.price - liquidity_to.price) * liquidity_from.day_volume) / 1000000)
